# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend Engineer Assignment API",
    "description": "This API allows for managing users, schools, and students in a system.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/user/login": {
      "post": {
        "summary": "Login a user",
        "tags": ["Users"],
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/getUser": {
      "get": {
        "summary": "Get details of the logged-in user",
        "tags": ["Users"],
        "operationId": "getUser",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/student/updateStudent/{studentId}": {
      "put": {
        "summary": "Update student",
        "tags": ["Students"],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67733e4b47cfbed0f400dac0"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "schoolId": {
                    "type": "string"
                  },
                  "classroomId": {
                    "type": "string"
                  }
                },
                "example": {
                  "firstName": "Lala",
                  "lastName": "Dolla",
                  "schoolId": "64abc1234def567890abcdef",
                  "classroomId": "64abc9876def123450abcdef"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/student/deleteStudent/{studentId}": {
      "delete": {
        "summary": "Delete student",
        "tags": ["Students"],
        "operationId": "deleteStudent",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67733f1ff094940daa1f2baa"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted"
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/classroom/getClassrooms": {
      "get": {
        "summary": "Get all classrooms",
        "tags": ["ClassRooms"],
        "operationId": "getClassrooms",
        "responses": {
          "200": {
            "description": "List of classrooms"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/classroom/createClassroom": {
      "post": {
        "summary": "Create a new classroom",
        "tags": ["ClassRooms"],
        
        "operationId": "createClassroom",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "resources": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isSmartClassroom": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "name": "Grade 3B",
                  "capacity": 25,
                  "resources": [
                    "table",
                    "fan"
                  ],
                  "isSmartClassroom": true
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Classroom created"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/classroom/updateClassroom/{classroomId}": {
      "put": {
        "summary": "Update classroom",
        "tags": ["ClassRooms"],
        "operationId": "updateClassroom",
        "parameters": [
          {
            "name": "classroomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "677340d5afe57a869bcbcfd5"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer"
                  },
                  "resources": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isSmartClassroom": {
                    "type": "boolean"
                  }
                },
                "example": {
                  "name": "Grade 10A",
                  "capacity": 100,
                  "schoolId": "640d1f44fc13ae1f35000002",
                  "resources": [
                    "table",
                    "fan"
                  ],
                  "isSmartClassroom": true
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Classroom updated"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/classroom/deleteClassroom/{classroomId}": {
      "delete": {
        "summary": "Delete classroom",
        "tags": ["ClassRooms"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "operationId": "deleteClassroom",
        "parameters": [
          {
            "name": "classroomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "677340d5afe57a869bcbcfd5"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Classroom deleted"
          },
          "404": {
            "description": "Classroom not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}